pyproject.toml
[tool.poetry]
name = "ecs-2"
version = "0.1.0"
description = ""
authors = ["tal.aizenkraft <tal.aizenkraft@tufin.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.104.1"
uvicorn = "^0.24.0.post1"
jinja2 = "^3.1.2"
black = "^23.11.0"
python-multipart = "^0.0.6"
motor = "^3.3.2"
fastapi-sessions = "^0.3.2"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
